KKK5ZR4 do not re-read and re-replay the log to answer a read
Every board-state request came through boot(), which read the whole log from
disk and materialized every event before deriving anything. On a twenty-person
decade that is 169 MB and 960,000 events — about eight seconds — for a board
that had not changed since the request a moment before.

boot() takes the log's signature first and skips the load and the replay when
it matches the last one. Only while live: a checkout in the past is
materialized by the time-travel path and a boot would throw it away, which
bootStateFromEventLog already refuses, and skipping must not quietly stand in
for that refusal. The signature is recorded after a successful boot, so a
failed one is retried rather than remembered as done.

The signature moves to lib/event/log-signature.ts, since the timeline cache and
this now share it — one definition of "has the log moved", with the reason the
read must come before the load written where both can see it.

A local write still grows the log, so the next request after an edit pays a
full reload. Reads are what this fixes.
